+++ site/src/components/SSHButton/SSHButton.tsx
@@ -1,6 +1,6 @@
-import Button from "@material-ui/core/Button"
 import Popover from "@material-ui/core/Popover"
 import { makeStyles } from "@material-ui/core/styles"
+import { SecondaryAgentButton } from "components/Resources/AgentButton"
 import { useRef, useState } from "react"
 import { CodeExample } from "../CodeExample/CodeExample"
 import { Stack } from "../Stack/Stack"
@@ -34,17 +34,14 @@ export const SSHButton: React.FC<React.PropsWithChildren<SSHButtonProps>> = ({
 
   return (
     <>
-      <Button
-        className={styles.button}
-        variant="outlined"
-        size="small"
+      <SecondaryAgentButton
         ref={anchorRef}
         onClick={() => {
           setIsOpen(true)
         }}
       >
         Connect SSH
-      </Button>
+      </SecondaryAgentButton>
       <Popover
         classes={{ paper: styles.popoverPaper }}
         id={id}
@@ -125,13 +122,5 @@ const useStyles = makeStyles((theme) => ({
 
   textHelper: {
     fontWeight: 400,
-  },
-
-  button: {
-    fontSize: 12,
-    fontWeight: 500,
-    height: theme.spacing(4),
-    minHeight: theme.spacing(4),
-    borderRadius: 4,
   },
 }))
