+++ site/src/components/VSCodeDesktopButton/VSCodeDesktopButton.tsx
@@ -1,8 +1,7 @@
-import { makeStyles } from "@material-ui/core/styles"
-import Button from "@material-ui/core/Button"
 import { getApiKey } from "api/api"
 import { VSCodeIcon } from "components/Icons/VSCodeIcon"
 import { FC, PropsWithChildren, useState } from "react"
+import { PrimaryAgentButton } from "components/Resources/AgentButton"
 
 export interface VSCodeDesktopButtonProps {
   userName: string
@@ -15,13 +14,11 @@ export const VSCodeDesktopButton: FC<
   PropsWithChildren<VSCodeDesktopButtonProps>
 > = ({ userName, workspaceName, agentName, folderPath }) => {
   const [loading, setLoading] = useState(false)
-  const styles = useStyles()
 
   return (
-    <Button
+    <PrimaryAgentButton
       startIcon={<VSCodeIcon />}
       disabled={loading}
-      className={styles.button}
       onClick={() => {
         setLoading(true)
         getApiKey()
@@ -50,31 +47,6 @@ export const VSCodeDesktopButton: FC<
       }}
     >
       VS Code Desktop
-    </Button>
+    </PrimaryAgentButton>
   )
 }
-
-const useStyles = makeStyles((theme) => ({
-  button: {
-    whiteSpace: "nowrap",
-    backgroundColor: theme.palette.background.default,
-    padding: theme.spacing(0, 3),
-    height: 44,
-    borderRadius: 6,
-
-    "&:hover": {
-      backgroundColor: `${theme.palette.background.paper} !important`,
-    },
-
-    "& .MuiButton-startIcon": {
-      marginRight: theme.spacing(1.5),
-      width: 16,
-      height: 16,
-
-      "& svg": {
-        width: "100%",
-        height: "100%",
-      },
-    },
-  },
-}))
